finalizando produto

+++ ProjetoEcommerce/Areas/Admin/Models/Produto.cs
@@ -9,7 +9,7 @@ namespace ProjetoEcommerce.Areas.Admin.Models
     {
         public int Id { get; set; }
         public string Nome { get; set; }
-        public double Preco { get; set; }
+        public decimal Preco { get; set; }
         public int Estoque { get; set; }
         public string Categoria{ get; set; }
     }
